+++ lib/src/main/java/com/diffplug/spotless/FilterByContentPatternFormatterStep.java
@@ -1,5 +1,5 @@
 /*
- * Copyright 2016-2021 DiffPlug
+ * Copyright 2016-2022 DiffPlug
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -16,8 +16,9 @@
 package com.diffplug.spotless;
 
 import java.io.File;
+import java.util.Collections;
+import java.util.List;
 import java.util.Objects;
-import java.util.regex.Matcher;
 import java.util.regex.Pattern;
 
 import javax.annotation.Nullable;
@@ -40,13 +41,21 @@ final class FilterByContentPatternFormatterStep implements FormatterStep {
 	public @Nullable String format(String raw, File file) throws Exception {
 		Objects.requireNonNull(raw, "raw");
 		Objects.requireNonNull(file, "file");
-
-		Matcher matcher = contentPattern.matcher(raw);
-
-		if (matcher.find()) {
+		if (contentPattern.matcher(raw).find()) {
 			return delegateStep.format(raw, file);
 		} else {
 			return raw;
+		}
+	}
+
+	@Override
+	public List<Lint> lint(String content, File file) throws Exception {
+		Objects.requireNonNull(content, "content");
+		Objects.requireNonNull(file, "file");
+		if (contentPattern.matcher(content).find()) {
+			return delegateStep.lint(content, file);
+		} else {
+			return Collections.emptyList();
 		}
 	}