+++ lib/src/main/java/com/diffplug/spotless/FilterByFileFormatterStep.java
@@ -1,5 +1,5 @@
 /*
- * Copyright 2016 DiffPlug
+ * Copyright 2016-2022 DiffPlug
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -16,6 +16,8 @@
 package com.diffplug.spotless;
 
 import java.io.File;
+import java.util.Collections;
+import java.util.List;
 import java.util.Objects;
 
 import javax.annotation.Nullable;
@@ -42,6 +44,17 @@ final class FilterByFileFormatterStep implements FormatterStep {
 			return delegateStep.format(raw, file);
 		} else {
 			return raw;
+		}
+	}
+
+	@Override
+	public List<Lint> lint(String content, File file) throws Exception {
+		Objects.requireNonNull(content, "content");
+		Objects.requireNonNull(file, "file");
+		if (filter.accept(file)) {
+			return delegateStep.lint(content, file);
+		} else {
+			return Collections.emptyList();
 		}
 	}